Split core of DownloadResponse back into a DownloadResponseTrait
I think that I misunderstood what was being asked originally regarding
working with a DownloadResponseTrait and a CsvResponse. Given that, this
change is being made to create a DownloadResponse but keep the core of
the functions in the trait. Not sure if this is the correct approach,
but it makes sense to me at the moment.