Rework the DownloadResponse and DownloadResponseTrait
The intent here is to simplify how the DownloadResponseTrait works,
primarily, and therefore, how the DownloadResponse class works. I've
tried to simplify the way it checks that the download headers aren't
being overridden, how the download headers are set, that it doesn't try
to do too much, nor be oversmart about it.